neuro_explorer_train/requirements.txt
absl-py==1.4.0
angles==1.9.13
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astunparse==1.6.3
attrs==23.1.0
beautifulsoup4==4.12.2
bondpy==1.8.6
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
comm==0.1.3
contourpy==1.1.0
coverage==7.2.7
cv-bridge==1.16.2
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
diagnostic-updater==1.11.0
distro==1.8.0
docutils==0.20.1
dynamic-reconfigure==1.7.3
exceptiongroup==1.1.2
execnet==2.0.2
executing==1.2.0
fastjsonschema==2.17.1
flake8==6.1.0
flake8-docstrings==1.7.0
flatbuffers==23.5.26
fonttools==4.40.0
fqdn==1.5.1
gast==0.4.0
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.0
h5py==3.9.0
idna==3.4
image-geometry==1.16.2
imageio==2.31.1
iniconfig==2.0.0
interactive-markers==1.12.0
isoduration==20.11.0
isort==5.12.0
jax==0.4.13
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
jsonpointer==2.4
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
keras==2.12.0
keras-unet-collection==0.1.13
kiwisolver==1.4.4
lazy_loader==0.3
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mccabe==0.7.0
message-filters==1.16.0
mistune==3.0.1
ml-dtypes==0.2.0
nest-asyncio==1.5.6
netifaces==0.11.0
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.8.0.74
opencv-python-headless==4.8.0.74
opt-einsum==3.3.0
overrides==7.3.1
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pilz-robot-programming==0.5.0
platformdirs==3.8.1
pluggy==1.2.0
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.0
pycparser==2.21
pydocstyle==6.3.0
pydot==1.4.2
pyflakes==3.1.0
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-json-logger==2.0.7
python-qt-binding==0.4.4
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
referencing==0.29.1
requests==2.31.0
requests-oauthlib==1.3.1
resource_retriever==1.12.7
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.8.8
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
Send2Trash==1.8.2
six==1.16.0
smclib==1.8.6
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.4.1
srdfdom==0.6.4
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow==2.12.1
tensorflow-addons==0.21.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.4.0
threadpoolctl==3.2.0
tifffile==2023.7.18
traitlets==5.9.0
typeguard==2.13.3
typing_extensions==4.5.0
tzdata==2023.3
urdfdom-py==0.4.6
uri-template==1.3.0
urllib3==1.26.16
validators==0.20.0
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.3.6
wrapt==1.14.1
xacro==1.14.16
zipp==3.15.0
